addUrl&deleteTel_JH
