+++ src/main/java/com/example/demo/store/controller/SpotController.java
@@ -1,6 +1,6 @@
 package com.example.demo.store.controller;
 
-import com.example.demo.store.dto.SpotDTO.SpotDTO;
+import com.example.demo.store.dto.SpotDTO;
 import com.example.demo.store.service.SpotService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -48,9 +48,22 @@ public class SpotController {
         return ResponseEntity.ok(spotList);
     }
 
+    // 사용자 서버 요청 시 Spot 리스트 반환하는 메서드
+    @GetMapping("/user/list")
+    public ResponseEntity<List<SpotDTO>> forUserList() {
+        log.info("----- Spot List Request -----");
+
+        List<SpotDTO> spotList = spotService.list();
+
+        log.info("Successfully retrieved Spot List");
+        log.info("Spot List: {}", spotList);
+        log.info("----- End of Spot List Request -----");
+
+        return ResponseEntity.ok(spotList);
+    }
+
     // Spot 추가
     @PostMapping("/add")
-    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<SpotDTO> add(@RequestBody SpotDTO spotDTO) {
         log.info("----- Spot Addition Request -----");
         log.info("Received SpotDTO: {}", spotDTO);
